backend/apps/organizations/tasks.py
"""
Celery tasks for organization operations.
"""
from celery import shared_task
import logging

logger = logging.getLogger(__name__)


@shared_task
def sync_github_repositories(organization_id=None):
    """
    Sync repositories from GitHub for organizations.
    """
    from .models import Organization, Repository
    from github import Github
    from django.conf import settings

    try:
        if organization_id:
            organizations = Organization.objects.filter(id=organization_id, is_active=True)
        else:
            organizations = Organization.objects.filter(
                is_active=True,
                github_org_id__isnull=False
            )

        for org in organizations:
            if not org.github_org_name:
                continue

            # TODO: Implement GitHub API sync
            # 1. Authenticate with GitHub App
            # 2. Fetch organization repositories
            # 3. Create/update Repository and Branch models
            # 4. Mark inactive repositories

            logger.info(f"Synced repositories for organization: {org.name}")

    except Exception as e:
        logger.error(f"Error syncing GitHub repositories: {e}")
        raise
