backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9
django-postgres-extra==2.0.8

# Authentication & Authorization
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1

# Caching & Pub/Sub
redis==5.0.1
django-redis==5.4.0
hiredis==2.3.2

# Object Storage (S3/MinIO)
boto3==1.34.22
django-storages==1.14.2

# Task Queue (for async operations)
celery==5.3.6
celery[redis]==5.3.6

# API Documentation
drf-spectacular==0.27.0

# Security & Rate Limiting
django-ratelimit==4.1.0
django-axes==6.1.1

# Monitoring & Logging
sentry-sdk==1.39.2
python-json-logger==2.0.7

# SARIF Processing
sarif-om==1.0.4

# GitHub Integration
PyGithub==2.1.1
cryptography==42.0.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
urllib3==2.1.0

# Development & Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==22.0.0
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0
django-stubs==4.2.7
djangorestframework-stubs==3.14.5

# WSGI Server
gunicorn==21.2.0
uvicorn==0.27.0
